feat: Move resume to about temporarily
